refactor(mobile): split runtime/hq into a directory module
