+++ apps/mobile/src/runtime/hq/hosts.ts
@@ -1,0 +1,19 @@
+import { z } from 'zod';
+import { HQ_URL, hqAuthClient } from './client';
+
+export const OnlineHostSchema = z.object({
+  hostId: z.string().min(1),
+  name: z.string().nullable(),
+  connectedAt: z.number(),
+  lastSeen: z.number(),
+});
+export type OnlineHost = z.infer<typeof OnlineHostSchema>;
+
+/** The account's hosts currently connected to the relay (`GET /tunnel/hosts`). */
+export async function fetchOnlineHosts(): Promise<OnlineHost[]> {
+  const { data, error } = await hqAuthClient.$fetch<unknown>(`${HQ_URL}/tunnel/hosts`, {});
+  if (error) throw new Error(`host discovery failed (${error.status})`);
+  const parsed = z.object({ hosts: z.array(OnlineHostSchema) }).safeParse(data);
+  if (!parsed.success) throw new Error('host list returned an unexpected shape');
+  return parsed.data.hosts;
+}
